Fix collection of kernel crash into /var/spool/crash
No kcrash was collected because the IsEnabled() test always failed:
we must do the kernel_collector->Enable() first.  I've not checked
if this might break the Override path (for local testing): it's more
important for the feature to work for real than in an artificial test.

BUG=837,5865
TEST=logging_KernelCrashServer or manual procedure

Signed-off-by: Hugh Dickins <hughd@chromium.org>

Review URL: http://codereview.chromium.org/3317007