Add the touch command to toolbox.
I wrote this to test my fix to support utime(2) system calls in the
sdcard fuse filesystem for stingray, and decided to finish sprucing
it up and make it part of toolbox.  In an effort to keep it small,
it doesn't accept dates a la touch, but just a time_t value.

Change-Id: I5dd011cd2e34d0cc605d6f40e46b96a8c949f194