Add BinPackArguments: false to .clang-format
Without this setting, arguments passed to a function can be packed in
more than one line, with more than one argument per line.

For example, the following code will be formated as this without the
BinPackArguments:

  CallMyFunction(
    with_more_than_one, argument_per_line,
    event_they_dont_fit_in_a_single_line);

On the other hand, with BinPackArguments: false, that code will be
split in:
  CallMyFunction(
    with_more_than_one,
    argument_per_line,
    event_they_dont_fit_in_a_single_line);

or all the arguments in the same line if they all fit.

Bug: None
TEST=None

Change-Id: I48627e66e2a841f89418c0763e7fe6c3e6ea7387